Add echo back to fix git duet initials alignment
On my last two projects I've observed that the visualization of the git duet initials will be broken on OSX machines that have run this setup project. Instead of showing the pair initials in one line (such as dn + to) it instead puts a newline after the + sign.

I know it didn't used to do this, and it looks like there was a change that broke it with this commit:
https://github.com/Pivotal/workstation-setup/commit/d8e4bfa99848f8e219d21aa85439ebcbc580a376

I'm not super great with bash, but reinserting the echo that wraps the git config command appears to resolve the issue.